Move some memos TradeTableModal->RaidPartyContext
